% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwp-package.R
\docType{package}
\name{dwp}
\alias{dwp-package}
\title{dwp: Access Stat-Xplore data on the UK benefits system}
\description{
The Stat-Xplore Open Data API is a JSON REST API, with the same data as
on the \href{https://stat-xplore.dwp.gov.uk/}{Stat-Xplore} online service. All
queries require the use of an API key. To set up an API key,
use \code{\link[=dwp_api_key]{dwp_api_key()}}. The API is free to use, but queries are rate limited.
To find the number of allowable queries per hour, and the number of
queries remaining, use \code{\link[=dwp_rate_limit]{dwp_rate_limit()}}.
}
\details{
Full documentation of the API is available
(here)\link{https://stat-xplore.dwp.gov.uk/webapi/online-help/Open-Data-API.html}.
}
