% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{dwp_get_data}
\alias{dwp_get_data}
\title{Submit and receive table queries}
\usage{
dwp_get_data(database, measures, row, column, wafer)
}
\arguments{
\item{database}{The database}

\item{measures}{The measure to use}

\item{row}{Row fields}

\item{column}{Column fields}

\item{wafer}{Wafer fields}
}
\value{
A tibble
}
\description{
Submit and receive table queries
}
\examples{
\dontrun{

x <- dwp_get_data(database = "str:database:ESA_Caseload",
                  measures = "str:count:ESA_Caseload:V_F_ESA",
                  row = "str:field:ESA_Caseload:V_F_ESA:ICDGP",
                  column = c("str:field:ESA_Caseload:V_F_ESA:CCSEX",
                             "str:field:ESA_Caseload:V_F_ESA:CTDURTN"),
                  wafer = "str:field:ESA_Caseload:V_F_ESA:IB_MIG")


}
}
