% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{dwp_rate_limit}
\alias{dwp_rate_limit}
\title{Rate Limits}
\usage{
dwp_rate_limit()
}
\value{
A tibble with one row and three columns:
\enumerate{
\item \code{remaining} The number of queries remaining from the limit
\item \code{reset} The date and time the rate limit resets, in POSIXct class
\item \code{limit} The maximum number of queries allowed within a given time span
}
}
\description{
Identify rate limits, remaining requests and request limit reset point. This
function does not accept any parameters
}
