% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{dwp_schema}
\alias{dwp_schema}
\title{Schema data}
\usage{
dwp_schema(id = NULL)
}
\arguments{
\item{id}{}
}
\value{
A tibble.
}
\description{
Schema data is the detail on the datasets and folders available
at the root level of Stat-Xplore. More details on the \code{schema} endpoint is
available in the
\href{https://stat-xplore.dwp.gov.uk/webapi/online-help/Open-Data-API-Schema.html}{API documentation}
}
